// resources/js/ClassInterfaces/Summoner.ts
export interface Summoner {
    profileIconId:       number;
    name:                string;
    puuid:               string;
    summonerLevel:       number;
    accountId:           string;
    id:                  string;
    revisionDate:        number;
    lastUpdate:          string;
    lastUpdateMatchlist: string;
}